requirements.txt
openai>=1.0.0
rich>=13.0.0
flask>=3.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
pillow>=10.0.0
mss>=9.0.0
duckduckgo-search>=7.0.0
jieba>=0.42.1
httpx>=0.27.0
watchdog>=3.0.0
numpy<2.0.0
